% Auto-generated: do not edit by hand
\name{''MessageLog}

\alias{''MessageLog}

\title{MessageLog component}

\description{

}

\usage{
''MessageLog(id=NULL, log=NULL, trigServLogUpdate=NULL)
}

\arguments{
\item{id}{Character. }

\item{log}{List of characters. }

\item{trigServLogUpdate}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

