% Auto-generated: do not edit by hand
\name{''NumDisplay}

\alias{''NumDisplay}

\title{NumDisplay component}

\description{

}

\usage{
''NumDisplay(number=NULL, id=NULL)
}

\arguments{
\item{number}{Numeric. }

\item{id}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

